requirements.txt
tensorflow==2.3
psutil==5.8.0
PyYAML==5.1.1
